Support ContentProvider-based servers
This is a major refactor to support both:
1. Building stoic into an app, exported as a ContentProvider (debuggable
   not required)
2. Injecting stoic using JVMTI (as it was done in the past.

The protocol is simplified, so that we no longer manually copy plugin
jars - we have a LoadPlugin message instead. This means that its
possible to fast-path connect directly from your laptop, provided that
stoic's server is already running, which is significantly faster (about
250ms when connecting to an emulator).

I also fixed the problem of stoic occasionally failing with `--restart`
/ `--start-if-needed`. The issue is that pidof was failing even though
it succeeded previously (and the same process was still running). I
worked around this by ensuring that we only call pidof once, and we
store the result.