Remove shell/rsync
Since the architecture was changed to not sync plugins to the device and
not support calling stoic on the device, rsync was over-kill. The slow
path is now a couple of hundred milliseconds faster, and (more
importantly) the codebase is simpler.

I'll miss the shell command, but it's unrelated to stoic's main purpose
- it belongs in a separate repo.